workout-api/pyproject.toml
[tool.poetry]
name = "workouttracker"
version = "0.1.0"
description = ""
authors = ["Luis Enrique Ramirez <lerf1010@outlook.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
alembic = "^1.13.1"
python = "^3.11"
SQLAlchemy = "^2.0.36"
aiosqlite = "^0.20.0"
pydantic = "^2.10.4"
bcrypt = "^4.2.1"
asyncpg = "^0.30.0"
fastapi = "^0.115.6"
pyjwt = "^2.10.1"
uvicorn = "^0.34.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
